fix test order
